+++ glue/viewers/histogram/qt/data_viewer.py
@@ -1,6 +1,5 @@
 from __future__ import absolute_import, division, print_function
 
-from glue.utils import nonpartial
 from glue.viewers.matplotlib.qt.toolbar import MatplotlibViewerToolbar
 from glue.core.edit_subset_mode import EditSubsetMode
 from glue.core.util import update_ticks
@@ -31,11 +30,11 @@ class HistogramViewer(MatplotlibDataViewer):
 
     def __init__(self, session, parent=None, state=None):
         super(HistogramViewer, self).__init__(session, parent, state=state)
-        self.state.add_callback('x_att', nonpartial(self._update_axes))
-        self.state.add_callback('x_log', nonpartial(self._update_axes))
-        self.state.add_callback('normalize', nonpartial(self._update_axes))
+        self.state.add_callback('x_att', self._update_axes)
+        self.state.add_callback('x_log', self._update_axes)
+        self.state.add_callback('normalize', self._update_axes)
 
-    def _update_axes(self):
+    def _update_axes(self, *args):
 
         if self.state.x_att is not None: